// src/app/models/signin-model.ts
import { IResponseModel } from "../shared/models/IResponseModel";
import { IUser } from "./IUserModel";

export interface SignInModel {
  email: string;
  password: string;
}

// Generated by https://quicktype.io

export interface SiginResponseModel extends IResponseModel {
  data: ISignIn;
}

export interface ISignIn {
  id: string;
  auth_token: string;
  expires_in: number;
  roles: any[];
  user: IUser;
  canLogin: boolean;
  isLockedOut: false;
  isNotAllowed: boolean;
  requiresTwoFactor: boolean;
}